% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visa_SimuData.R
\docType{data}
\name{visa_SimuData}
\alias{visa_SimuData}
\title{visa_SimuData}
\format{
The dataset visa_SimuData contains n = 50 samples with p = 10 covariates and treatment variable
\describe{
\item{y}{the response}
\item{x}{the covariates}
\item{a}{the treatment received}
}
}
\usage{
visa_SimuData
}
\description{
An Example of Simulated Data for visa
}
\keyword{datasets}
